requirements.txt
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
geopandas>=0.14.0
psycopg2-binary>=2.9.0
shapely>=2.0.0
matplotlib>=3.7.0
folium>=0.15.0
sqlalchemy>=2.0.0
GeoAlchemy2>=0.14.0
python-dotenv>=1.0.0
pyproj>=3.5.0
fiona>=1.9.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
